it seems bitfields don't work in Linux kernel (anymore?!) so fixed it

+++ kipc/pcn_kmsg_test.h
@@ -42,7 +42,7 @@ struct pcn_kmsg_test_args {
 struct pcn_kmsg_test_message {
 	struct pcn_kmsg_hdr hdr;
 	enum pcn_kmsg_test_op op;
-	int src_cpu, dest_cpu; // needed to support clustering
+	short src_cpu, dest_cpu; // needed to support clustering
 	unsigned long batch_seqnum;
 	unsigned long batch_size;
 	unsigned long ts1, ts2, ts3, ts4, ts5;
